+++ app/src/main/java/com/nsd/talk/data/RetrofitInstance.kt
@@ -1,7 +1,5 @@
-package com.nsd.talk
+package com.nsd.talk.data
 
-import com.google.gson.Gson
-import okhttp3.OkHttpClient
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
