+++ app/src/main/java/com/nsd/talk/ui/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.nsd.talk
+package com.nsd.talk.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.nsd.talk.R
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
